+++ src/compas/rpc/proxy.py
@@ -353,12 +353,12 @@ if __name__ == "__main__":
     import compas
 
     from compas.datastructures import Mesh
-    # from compas.plotters import MeshPlotter
-    from compas_rhino.artists import MeshArtist
+    from compas.plotters import MeshPlotter
+    # from compas_rhino.artists import MeshArtist
 
     from compas.rpc import Proxy
 
-    numerical = Proxy('compas.numerical', python='pythonw')
+    numerical = Proxy('compas.numerical')
 
     print(numerical.python)
     print(numerical.address)
@@ -392,14 +392,14 @@ if __name__ == "__main__":
         attr['f'] = f[index][0]
         attr['l'] = l[index][0]
 
-    # plotter = MeshPlotter(mesh, figsize=(10, 7))
-    # plotter.draw_vertices()
-    # plotter.draw_faces()
-    # plotter.draw_edges()
-    # plotter.show()
+    plotter = MeshPlotter(mesh, figsize=(10, 7))
+    plotter.draw_vertices()
+    plotter.draw_faces()
+    plotter.draw_edges()
+    plotter.show()
 
-    artist = MeshArtist(mesh)
-    artist.draw_vertices()
-    artist.draw_edges()
-    artist.draw_faces()
-    artist.redraw()
+    # artist = MeshArtist(mesh)
+    # artist.draw_vertices()
+    # artist.draw_edges()
+    # artist.draw_faces()
+    # artist.redraw()