+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- `xscript.py` from `compas.utilities`
+- `mixing.py` from `compas.utilities`
+- `singleton.py` from `compas.utilities`
+
 
 ## [0.5.1] 2019-03-25